Create & Update user + Role Middleware

+++ backend/routes/api.php
@@ -16,10 +16,11 @@ Route::middleware('auth:sanctum')->group(function () {
      * -----------------------------------------
      */
     Route::prefix('users')->group(function () {
-        Route::post('', [UserController::class, 'create']);
+        Route::post('', [UserController::class, 'create'])->middleware('role:advanced');
         Route::get('', [UserController::class, 'current']);
+        Route::get('list', [UserController::class, 'list'])->middleware('role:advanced');
         Route::patch('{id}', [UserController::class, 'updatePassword']);
-        Route::delete('{id}', [UserController::class, 'remove']);
+        Route::delete('{id}', [UserController::class, 'remove'])->middleware('role:advanced');
     });
